fix one-to-many problem

+++ crm-microservice/src/main/java/com/example/crm/entity/Address.java
@@ -16,7 +16,7 @@ import org.hibernate.annotations.DynamicUpdate;
 public class Address {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	private String city;
 	private String country;
 	private String line;
@@ -27,11 +27,11 @@ public class Address {
 	public Address() {
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -79,7 +79,7 @@ public class Address {
 	public int hashCode() {
 		final int prime = 31;
 		int result = 1;
-		result = prime * result + id;
+		result = prime * result + id.intValue();
 		return result;
 	}
